+++ R/Manningtrap.R
@@ -188,6 +188,7 @@
 #' library(iemisc)
 #' library(iemiscdata)
 #' # Exercise 4.1 from Sturm (page 153)
+#' 
 #' Manningtrap(Q = 3000, b = 40, m = 3, Sf = 0.002, n = 0.025, units = "Eng")
 #' # Q = 3000 cfs, b = 40 ft, m = 3, Sf = 0.002 ft/ft, n = 0.025,
 #' # units = English units
@@ -201,11 +202,14 @@
 #' # Use the minimum Manning's n value for 1) Natural streams - minor streams
 #' # (top width at floodstage < 100 ft), 2) Lined or Constructed Channels,
 #' # 3) Concrete and 4) float finish.
+#'
 #' data(nchannel)
 #'
 #' nlocation <- grep("float finish",
 #' nchannel$"Type of Channel and Description")
+#'
 #' n <- nchannel[nlocation, 3][1] # 3 for column 3 - Normal n
+#'
 #' Manningtrap(y = 1.5, b = 3, m = 3/2, Sf = 0.1/100, n = n, units = "SI")
 #' # y = 1.5 m, b = 3 m, m = 3/2, Sf = 0.1/100 m/m, n = 0.023, units = SI
 #' # units