+++ R/Manningtri.R
@@ -188,11 +188,14 @@
 #' # Use the normal Manning's n value for 1) Natural streams - minor streams
 #' # (top width at floodstage < 100 ft), 2) Excavated or Dredged Channels, 3)
 #' # Earth, straight, and uniform, 4) clean, recently completed.
+#'
 #' data(nchannel)
 #'
 #' nlocation <- grep("clean, recently completed",
 #' nchannel$"Type of Channel and Description")
+#'
 #' n <- nchannel[nlocation, 3] # 3 for column 3 - Normal n
+#'
 #' Manningtri(Q = 0.68, m = 1.5, Sf = 0.0023, n = n, units = "Eng")
 #' # Q = 0.68 cfs, m = 1.5, Sf = 0.002 ft/ft, n = 0.05, units = English units
 #' # This will solve for y since it is missing and y will be in ft