+++ NEWS.md
@@ -1,11 +1,18 @@
+# iemisc 0.9.1
+
+* Added more examples to the README.md
+* Revised the examples in these functions: `Manningtri`, `Manningtrap`, `Manningrect`, `Manningpara`, `Manningcirc`
+* Updated the Open Channel Flow problems vignette
+
+
 # iemisc 0.9.0
 
 * Updated these functions: `Manningtri`, `Manningtrap`, `Manningrect`, `Manningpara`, `Manningcirc`
-* Added vignette for Open Channel Flow problems
+* Added an Open Channel Flow problems vignette
 * Added `install.load` as a suggested R package
 
 
-# iemisc 0.5.1
+# iemisc 0.5.2
 
 * Updated these functions: `size`, `righttri`, `Manningtri`, `Manningtrap`, `Manningrect`, `Manningpara`, `Manningcirc`